Let errors occur when copying the DPU CNI
It is most likely that we will in the future include a command line
argument for the CNI path since different cluster types will have
different CNI locations.

For now remove the check for "writeable()" and simplify the code to just
attempt to copy and make executable. Errors are logged and not act upon.

The writeable() code also had a bug where it would check for the file and
not the directory.

Signed-off-by: William Zhao <wizhao@redhat.com>