platform: implement IsDpu() based on detectDpuPlatform()
IsDpu() does almost the same as detectDpuPlatform(). Implement it based
on the latter.

IsDpu() iterates over the vendor detectors, and calls IsDpuPlatform() on
each. Previously, the search would finish when either the first error or
the first successful detection happened. That seems bad, because then
the result depends on the order in which detectors are listed. In
contrast, detectDpuPlatform() will try all detectors, collect the errors
and return if any errors happened.

Signed-off-by: Thomas Haller <thaller@redhat.com>