vendorplugin: render Command/Args in daemonset manifest
We could require that all VSP plugin images have the same
"command"/"args" to execute.

However, the vendor plugin may need a way to pass specific arguments to
the plugin. So there anyway needs to be a way to customize the
"command"/"args" of the deamonset. At this point, just make it fully
configurable.

While at it, for Intel IPU the command line arguments seem no longer
necessary. Drop most of them, except "-v=debug". See [1].

[1] https://github.com/intel/ipu-opi-plugins/commit/5d07f00942448bd02dbd54dd0207696a537f64e6

Signed-off-by: Thomas Haller <thaller@redhat.com>