Makefile: directly call scripts instead of using variables
These variables serve no purpose as they are only used once. Drop them,
and call the scripts directly.

This also fixes some of the unnecessary inconsistencies like

  PREPARE_SCRIPT = hack/prepare.sh
  MAKE_E2E_FAST = hack/deploy_fast.sh

  prepare-e2e-test:
      bash $(PREPARE_SCRIPT)
  fast_e2e_test: prepare-e2e-test
      bash $(MAKE_E2E_FAST)

where the variable names, the script names and the Make targets have no
obvious resemblance (with the change, the Make targets are still
inconsistent with the underlying scripts).

Also, the scripts are executable and have a shebang. They choose their
interpreter themselves and the Makefile should not assume that this is
always "bash".