Makefile: disable "deploy_tft_tests" step from "e2e-test" for now
It is currently not working yet. Disable it for now, otherwise the
"e2e-test" target will fail.

Note that currently there is a bug in "deploy_tft_tests" target, so
currently it is also not called. So this change has little effect.
However, it will have effect, when that bug is fixed (and the target
starts to fail for real).