hack: use local temporary directory instead of shared one in /tmp
Running some of the make targets in parallel is possibly not going to
work. That is undesirable, bug generally hard to avoid.

What however should work, is that two separate checkouts of the project
can build and test in parallel. Whether that will work is unclear,
however it would require that they don't share a temporary directory in
/tmp. Instead, create the temporary directory in the current checkout.
At least this aspect should not break parallel operations (for distinct
checkouts).