daemon: Crash if vsp init fails
In the case that the vsp is not able to be started, the daemon should
crash and retry.

Currently there is a race condition when deploying the vsp from the
operator where the vsp may fail to come up by the time "start" is
called. This patch should address this race by retrying the daemon pod
deployment.

It makes sense to crash anyways, since the daemon on its own is useless
without a running vsp, and we want to be able to easily see that an
error has occured (i.e. daemon ends up in crashLoopBackoff if the vsp is
never deployed).

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>