hack: Track the latest IPU sha we have validated
We are currently just building from tip of vsp, which can make it hard
to attribute failures to changes that make occur / determine when fixes
have properly been applied in the vsp.

Lets start tracking the current sha of the vsp that we are building.
Future changes to the vsp will need to be "validated" by pushing them
through the CI before we will encorporate those changes into the stable
solution.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>