hack: remove wa for IIC-364
In theory, we have addressed the issue in IIC-364 that caused the
connection from host to acc to be lost on host reboot.

Lets remove this workaround to validate this change has been applied.

Note: we also will remove the reboots we were issuing in the subsequent
bump to CDA, which means the issue may still be present but masked.

TODO: Add additional stress-testing to validate reboot behavior on
vendors platforms

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>