deviceHandler: Handle both net dev and pci IDs
Currently, we do not specify whether the deviceID provided by vendors
implmentation of GetDevices() should confirm to pci addresses, netdevs,
or anything else. This is just a blob that allows us to detect the
devices.

As a result we need to handle the case netdevs are provided. Internally
the device plugin expects all devies in PCI address format

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>