webhook: generate webhook using operator-sdk
This commit is fully generated without manual changes.

  $ make operator-sdk && \
    ./bin/operator-sdk create webhook --group config --version v1 --kind DpuOperatorConfig --defaulting --programmatic-validation && \
    make vendor && \
    make manifests && \
    (cd api/ && go get github.com/onsi/ginkgo/v2@v2.20.2) && \
    make vendor && \
    make manifests && \
    make generate && \
    sed $'33i \\\tconfigv1 "github.com/openshift/dpu-operator/api/v1"' cmd/main.go -i

Note that here we also generate the defaulting/mutating webhook. We
don't actually need that one (for now). But we generate it here, then
the next commit will revert that change. The benefit is to have a
explicit commit that clearly shows the mutating webhook, which later can
be reverted if we need it.

https://sdk.operatorframework.io/docs/building-operators/golang/webhook/