webhook: enable Service Serving Certificate (service.beta.openshift.io) for webhooks
We need to have suitable SSL certifictes for the webhook. Inject them
via Service Serving certificate ([1], [2]).

Note that the generated code suggests to use [CERTMANAGER]. But that
requires having the cert-manager operator, which is not installed by
default in openshift ([3]).

Note that sriov-network-operator also defaults to Service Serving
Certificate, however, I think it also allows to use cert-manager. This
commit does not enable that for dpu-operator.

Also run `make bundle` and commit the changed files.

[1] https://docs.openshift.com/container-platform/4.16/security/certificates/service-serving-certificate.html
[2] https://github.com/openshift/service-ca-operator/tree/master
[3] https://docs.openshift.com/container-platform/4.16/security/cert_manager_operator/cert-manager-operator-install.html