webhook: validate Name of DpuOperatorConfig
It makes no sense to have more than one DpuOperatorConfig instance.
Enforce uniqueness, by ensuring that the name of the config is
"dpu-operator-config".

Note that "sriov-network-operator" has a similar check ([1]). One
difference is that the name is enforced to be "default". That doesn't
seem a good name to me (because "default" makes it sound as if you could
still create a non-default instance, which you cannot). Also, the name
"dpu-operator-config" is already established and used.

[1] https://github.com/openshift/sriov-network-operator/blob/45f22daa4ca8fb10f64e280d0aa2b7925da717d7/pkg/webhook/validate.go#L42

https://issues.redhat.com/browse/IIC-317