internal/utils/cluster_environment.go: Change the flavour focus away from microshift vs openshift
The reason why we had the flavours as Openshift vs Microshift is that we
assumed that if the DPU Operator is running on Openshift then it runs an
immutable operating system, which in this case is CoreOS. Like wise if
the operator is running on Microshift we assumed that the node is
running classic RHEL (mutable). The assumptions are not wrong in our
current use cases but microshift's target audience seems to be for
immutable flavors of RHEL, so we should make this more robust by
checking whether a system is Classic RHEL or if its an OSTREE-Based
RHEL, since it gives us a better way to know if parts of the filesystem
is writeble or if binaries can even be installed on it by the Operator.
This would also add compatibility with running Microshift on RHEL Image
Mode, RHEL for edge or RHEL CoreOS on the DPU.