webhook: generate webhook using operator-sdk but without defaulting/mutating
For now, we only need a validating and not a mutating webhook.

The parent commit generated both a validating and a mutating commit.
This commit on top of that and undoes the parts that are about enabling
the mutating webhook.

It's done this way, so we first generate both kinds of webhooks, and
then undo part of it. The benefit is that if we ever want to add the
mutating webhook, we can git-revert this commit. We also clearer see the
difference by doing it this way.

This commit has no manual changes, it is fully generated by running
the following command on the parent commit.

      $ make operator-sdk && \
        ./bin/operator-sdk create webhook --group config --version v1 --kind DpuOperatorConfig --programmatic-validation && \
        make vendor && \
        make manifests && \
        (cd api/ && go get github.com/onsi/ginkgo/v2@v2.20.2) && \
        make vendor && \
        make manifests && \
        make generate