controller/test: adjust controller-test for new webhook
Theoretically, we just setup the webhook and can use the webhook in our
controller test. Somehow related to WebhookInstallOptions it should
work. Except, I don't get it to work.

Instead, there webhook is not running in this unit test. We however have
them in the "Webhook Suite", so there are unit tests about the webhook
there.

Instead, adjust the controller test for the (lack of) webhook changes.
Note that this doesn't actually check the webhook, since it's not
running in this setup. It rather tests, that the webhook is not running.
If we ever make the webhook working for dpuoperatorconfig_controller_test
test, we will need to adjust this. That is intentional.