controller/test: adjust controller-test for new webhook
Theoretically, we just setup the webhook and can use the webhook in our
controller test. Somehow related to WebhookInstallOptions it should
work. Except, I don't get it to work.

Instead, there are not webhook in this unit test. We however have them
in the "Webhook Suite", so there is that.

Instead, adjust the controller test for the webhook changes. Note that
this doesn't actually check the webhook, since it's not running in this
setup.