marvell: set IPv6 link local addresses directly
Don't use NetworkManager to configure IPv6 link local addresses. I think
that is not desirable for serval reasons:

 - it adds a dependency to the VSP to only work on CoreOS hosts (or
   otherwise hosts that happen to use NetworkManager).

 - also, just calling `nmcli connection add` is not sufficient. We would
   at least need to handle the case where our profile is already created
   and make sure that the desired profile gets activated. Getting this
   right would be more complicated.

 - while we use nsenter to call nmcli in PID1's namespaces, the VSP
   plugin runs as non-root user, hence the request is rejected:

    # oc exec -ti pod/vsp-97tbs -- nsenter -t 1 -m -u -n -i -- nmcli connection add type ethernet
    Error: Failed to add 'ethernet' connection: Insufficient privileges
    command terminated with exit code 4

   This probably could be fixed, but is also effort.

 - NetworkManager is about user-facing, persisted, network configuration,
   so that other tools can also do something with that configuration.
   Here we have a tool that just wants to setup the link in a certain way.
   It seems unnecessary (undesirable?) to bother NetworkManager with this
   and expose connection profiles to the user.

Instead, use iproute2 to directly configure the link and addresses.

While at it, also configure fixed link local addresses fe80::1/64 and
fe80::2/64 on the DPU and the host, respectively. That way, we know the
IP addresses we can expect.

Also, the ping command

  nsenter -t 1 -m -u -n -i -- /usr/bin/ping6 ...

was also not working, because the host's CoreOS installs ping6 at /usr/sbin.
It's also unnecessary to use nsenter for this, because the VSP container
contains iputils package, so it can just use the one it has installed at
/usr/sbin/ping6 -- it's in the right netns already.