Remove left over -a from build from manager
All go build commands should be without a rebuild of all packages
as that takes a long time, and there is no real to do so. Currently,
most of the containers don't do a full rebuild, and the manager
was the only one that had a -a. Note also that Dockerfile.CNI.rhel
also uses a -a, but that container isn't really used.

Signed-off-by: Balazs Nemeth <bnemeth@redhat.com>