marvell: run VSP as root and set interface unmanaged in NetworkManager
In common cases, on the host NetworkManager is running and by
default the Marvell PF/VFs are managed. Also, by default NetworkManager
creates "Wired connection NUMBER" profiles for unknown ethernet
interfaces. While this is configurable, we get interference from
NetworkManager by default.

Issue a `nmcli device set $IFNAME managed no`, which makes
NetworkManager give up on that interface. This configuration is only
persisted in /run and lost after reboot, but that is fine for us. Next
time the VSP runs again, it will do it again.

To issue the nmcli command we enter the mount namespace of PID 1. There
we find nmcli. However, we also must run as User 0 (root), otherwise the
request will be denied. Run the pod as root user for that reason.