hack: call ./cda.py directly without explicit python interpreter
For one, it is well understood that we rely on having a suitable venv
environment loaded. That is even necessary, because CDA requires Python
3.11+ and we usually run on RHEL 9 (where the system python is Python
3.9).

If we load a venv, we should not longer call `python3.11 cda.py`.
Instead, we should either use `python3 cda.py` or just `python cda.py`.
The venv ensures that both those binaries exist in the path.

The benefit is, when we tomorrow upgrade to another Python version, then
we only need to change that at one place. Also, we don't need to
redunantly specify the python version at various places.

But actually, go one step further. "cda.py" is an executable and has a
suitable shebang of `#!/usr/bin/env python3`. Obviously it chose a
shebang that is suitable for our purposes here. Let's rely on that and
don't let callers second-guess the right interpreter. So instead of "python
cda.py" just call "./cda.py".

Signed-off-by: Thomas Haller <thaller@redhat.com>