hack: bump cda to partial remove IIC-427 WA
With the latest IPU Plugin changes, the user should not have to manually
install and run this ipu port 1 set up script.

However, there is still an edge case where the ACC will fail to come up
with networking on P1 enabled, which is after installing a new RHEL iso.
As a result, leave the retry for the time being.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>