cluster-configs: remove the bmc.user from "config-dpu.yaml"
In CDA, internally the bmc.user already defaults to "root". So setting
this had no real effect. Using "root" is suitable for the IPU clusters
that we have.

However, we want that this same config also works against Marvell DPUs.
In such clusters, the BMC is the x86_64 host that contains the DPU. We
usually expect that host to be an OCP node and run CoreOS. As such, we
assume that we can ssh into the host using "core" user and public key
authentication.

Note that currently the bmc.user cannot be templated from the google
spreadsheet. To make "hack/cluster-configs/config-dpu.yaml" work with
Marvell DPUs, CDA overwrites the "bmc.user" setting when deploying a
Marvell DPU and always uses "core".

Drop the bmc.user setting there, because it is confusing when CDA (for
Marvell DPU) doesn't actually use the value.