marvell: accept existing interface in createVethPair()
This should should be idempotent and accept that the interface might already
exist. Especially, since createVethPair() can be called multiple times.

  I0214 10:43:39.226702       1 main.go:218] Received Init() request: DpuMode: true
  ...
  I0214 10:43:39.542460       1 main.go:161] Creating Veth Pairs: 2
  ...
  I0214 10:44:06.202503       1 main.go:218] Received Init() request: DpuMode: true
  ...
  I0214 10:44:06.517237       1 main.go:161] Creating Veth Pairs: 2
  E0214 10:44:06.517348       1 main.go:165] Error occurred in creating Veth Pair: file exists
  I0214 10:44:06.719772       1 main.go:147] Deleted Veth Pair: nf_interface0
  I0214 10:44:06.919443       1 main.go:147] Deleted Veth Pair: nf_interface1
  E0214 10:44:06.919460       1 main.go:227] Error occurred in configuring Network Interface: file exists

It was indicated that the VSP should recognize when it already
called createVethPair() earlier and skip that. That may be a solution,
however, it does not conflict with createVethPair() sensibly handling
the case where the interfaces were already created and making the
function idempotent.