Set vsp owner reference to dpuOperatorConfig
We want the vsp daemonset to be owned by some resource such that the vsp
will be garbage collected properly.

Ideally this would be the daemon or dpu-daemon daemonset, but kubernetes
does not allow pods or daemonsets to own other controllers.

A good step in the right direction is to have the vsp owned by the same
custom resource that owns the dpu daemon.

A potenital next step might be to add an additional manager for
coordinating deployment of the vsp.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>