tests: use default name for DpuOperatorConfig
Soon our validating webhook will enforce that the name of
the DpuOperatorConfig is "dpu-operator-config". One point
of that is to enforce uniqueness of the config.

In the first step, these tests here will not run the validating webhook.
So nothing enforces a certain name for them.

However, it is still confusing that the DpuOperatorConfig in all other
cases will have a certain name.

Here, we don't want to test behavior for what happens when we have a
different name. Those tests just choose some name. Adjust the test so
they use the default name.