webhook: fix webhook after generating code
operator-sdk generates the webhook code that targets
sigs.k8s.io/controller-runtime v0.19 or older.

As such, it contains webhook.Validator and webhook.Defaulter.

This API was removed in v0.20, see [1].

To make this work, we patched the code out while generating it. But now
we have to fix it. Fix it now for the new API.

[1] https://github.com/kubernetes-sigs/controller-runtime/pull/2877