tft: fix script to only evaluate latest test run
On Jenkins, the workspace gets wiped on each run, so the "ft-logs"
directory is empty and there is no problem.

However, it's also useful to manually access the cluster and run `make
traffic-flow-tests` (or a full `make e2e-test`). In that case, the
result files pile up and "ft-logs/*.json" also evaluates past result
files.

Fix that by using the "--output-base" parameter, which exists for this
case.