vsp: remove unused volume from "99.vendor-plugin-daemonset.yaml"
volume "host-libmodules" is not used, and likely it should
not be used to not mix core directories of the container
with directories from the host. There is already host-root
path ("/host") for that.

While at it, add a tailing newline to the file. Text files without
trailing newline look odd in editors.