tft: make traffic_flow_tests more configurable to use with any cluster
`make e2e-test` and `make traffic-flow-tests` call
`hack/traffic_flow_tests.sh`, which runs the tests. It uses a
configuration file and detects part of the configuration.

Most notably it assumes to find the KUBECONFIG files at a certain place
and find a worker node with a certain naming pattern in `oc get node`.

This is useful for when we run the script as part of the `e2e-test,
so that the cluster is setup in a certain way that works for the traffic
flow tests script.

However, it would be useful to be able to run the tests with any
cluster, without having to follow the naming choices.

As "hack/traffic_flow_tests.sh" is quite trivial, you could just fork
the script and modify it to your needs. But optimally, we want a way to
run the test from this repository (mostly) as they are, without forking
parts.

Make that possible by honoring the environment variables

  - $TFT_KUBECONFIG
  - $TFT_KUBECONFIG_INFRA
  - $TFT_WORKER

with their obvious(?) meaning. Set these to overwrite what is being
tested.