prepare-venv: add file locking to avoid concurrent run
We have broken "/tmp/tft-venv" environments. It's not clear how that can
happen, maybe "prepare-venv.sh" runs in parallel (unlikely as it seems).

Note that setting up a python venv cannot be done concurrently.

Avoid that potential problem by adding a file lock and see whether that
avoids the corruption.