marvell: add "vendor.sh" script for including "pcie_ep_octeon_target"
During container build, we are not able to pull data from the internet.
Instead, vendor in the files that we will need.

Add a "vendor.sh" script. You can re-run it to re-fetch the sources.

The import of these sources should not be done manually, instead it
should be reproducible by running this script. You can also re-run
the script just to check whether what we have commited to git is the
same as what we would pull again.

Note that the script hard codes the git commit sha. So we will always
pull the exact same version (or fail). Obviously, when upgrading to a
newer version, the SHA sums should be adjusted and the script re-run.