marvell: avoid setting link down to workaround kernel issue
Due to RHEL-90248, it seems to help to keep the SDP interfaces up at all times.
The exact issue is not yet understood, but for now, avoid doing that while
setting addrgenmode.

Changing the addrgenmode only takes effect (and causes kernel to regenerate
IPv6 link local addresses) when bringing the link up. If it was up already, we
need to bring it down first. This is what the code previously tried to do.

Let's avoid that. If the addrgenmode is already eui64, leave it.

Also, in the meantime we use a fixed link local address IPv6AddrDpu.
Technically, we don't even need the EUI64 link local address.