e2e-test: add assertion that a SFC pod requests two "openshift.io/dpu" resources
We are interested whether resources are no leaked. We can see "Allocated
resources" in `oc describe node/NODE`. However, that value is only
computed client-side based on the pods that are running.

So, we really can only check, that the pod requests the expected number
of resources. That is currently hard coded at [1]. So the test merely
checks for that.

Maybe in the future, some VSPs won't require any physical resources for
SFC pods. E.g. if we use veth pairs that are not limited. In that case,
the test will need adjustment. But for now, merely check what we expect
to be the case.

[1] https://github.com/openshift/dpu-operator/blob/d1ec588f8960ea2017f77545e58f19320cc8ca38/internal/daemon/sfc-reconciler/sfc.go#L55