Add a new pcie field to the Init request from Daemon to VSP
The new dpu_identifier field is used by the daemon to tell the VSP
which DPU device is the VSP supposed to handle. Since the daemon
running on the host is responsible for discovery of devices on the
host. It should be responsible for passing this information to the
VSP such that it knows which device it is responsible for.

The purpose of the VSP is not to manage all devices, but handle the API
for managing one device, hense passing in dpu_identifier field is
appropriate small step forward to eventually supporting multiple
DPU devices on the same node.

Signed-off-by: William Zhao <wizhao@redhat.com>