Add Intel NetSec Accelerator Support with DPU Operator
This commit adds the intial PTL support for the DPU Operator. The goal
of this commit is to have the daemon properly detect the Intel NetSec
Accelerator. Then deploy the VSP properly. This commit deploys the
operator in a 1 cluster mode. Which means all daemons and vsps run in
1 cluster for the host and the Intel NetSec Accelerator. There was
changes to the taskfiles to achieve this.

Also remove duplication of code between Marvell and Intel NetSec
by introducing the new vspnetutils and using platform and afero
where appropriate.

For PTL we must use the PCIe Serial Number for the DPU Identifier. This
is because the PTL has 2 PFs.

Signed-off-by: William Zhao <wizhao@redhat.com>