testutils: extend EventuallyPodIsRunning() to wait additional time
Extend the code. If the pod is not running within timeout, then wait for
an additional timeout to see whether it would have been ready, if we
only waited longer. That additional waiting does not recover the
failure, the result is always and still an assertion failure. What we
get however is the information whether a longer timeout would have
helped.