marvell: fix Init() before VFs are created
With commit 1525e3a52f86 ('The StartVsp() function must be called first
prior to do any other API'), we call Init() before SetNumVfs().

Thus, usually, there are no VFs at the time when Init() is called. That
then leads to failure

   ERROR   Daemon  Failed to start VSP in sideManager      {"error": "failed calling VSP Start(): failed to start serving after 10s: rpc error: code = Unknown desc = no devices found with given Vendor ID and Device ID"}

This is because Init() does

        VfsPCI, err := mrvlutils.GetAllVfsByDeviceID(HostVFDeviceID)
        if err != nil {
            return nil, err
        }

Note that mrvlutils.GetAllVfsByDeviceID() returns an error, if no VFs
exist yet. We usually expect that no VFs exist yet at that point, and
must accept that gracefully.

Also, later, when we actually create the VFs, we must initialize the
deviceStore with the created devices.

Fixes: 1525e3a52f86 ('The StartVsp() function must be called first prior to do any other API')