testutils: add TryEventually()/AssertEventually() helpers
TryEventually() polls for a condition with timeout. It either
returns nil if the condition is met or the last error on timeout.

AssertEventually() also polls for a condition. On success it returns.
On failure, it will poll additional time to see whether waiting longer
would resolve the problem. In any case, the function will fail the
assertion at that point. The point is to collect more information
about why we are going to fail (i.e., would a longer timeout have
prevented the failure?).