build: bump required go version from 1.23.6 to 1.24.3
We now use new base container images to build our go code. Those images
bring go version 1.24.3.

The go directive in the "go.mod" file indicates the minimum version that
is required to build the module. So it's not strictly necessary to bump
this version, only because we use a newer go compiler.

Still do it for unspecified reasons.

https://go.dev/ref/mod#go-mod-file-go