prepare-venv: fix checking for valid ktoolbox package in venv
ktoolbox is usually

  https://raw.githubusercontent.com/thom311/ktoolbox/85ed9ba7ff346f20a02c78beffba9df08c67729a/wheels/155.6601dd6cdf03/ktoolbox-0.8.0-py3-none-any.whl

or it could be

  git+https://github.com/thom311/ktoolbox@6601dd6cdf031f2f0daeb6ebe333cd6e0740e83a

This requires special handling.

This does not work for all URL in general, because the URL could be
anything and not necessarily follow the pattern that we use to extract
the package name. Still, this works for the ktoolbox URLs, and maybe it
would work for others.

Otherwise, the venv always appears broken, and we re-create it. Actually
twice during one `task e2e-test` run.