Workaround issue with e810/e823 interfaces are not VLAN aware
Intel NetSec Acclerator e810 interfaces are not VLAN-aware in VEB mode,
thus packets can be forwarded to MAC addresses across VLAN boundaries.
This violates the IEEE802.1Q forwarding rules.

This issue is tracked here: https://issues.redhat.com/browse/RHEL-104927

To workaround this we need to set the host side to vepa mode. This would
bypass the loopback on the bridge and the bridging features entirely.
Intel is working on a fix to make the ethernet controllers become proper
VLAN aware bridges.

Signed-off-by: William Zhao <wizhao@redhat.com>