marvell: update octep_cp_agent to 25.04.0 for new firmware
So far we use Firmware

  Firmware Version: 2025-01-30 22:07:22
  EBF Version: 12.25.01, Branch: /home/fae/PANW-LIO3/SDK122501/work/cn10ka-pcie-ep-release-output/build/marvell-external-fw-SDK12.25.01/firmware/ebf, Built: Thu, 30 Jan 2025 22:06:12 +0000

with octep_cp_agent version 25.03.0 (commit aa84a2331f76).

We will want to update to

  Firmware Version: 2025-07-30 15:48:38
  EBF Version: 12.25.06, Branch: /sdk/SDK12.25.06/cn10ka-vpp-release-output/build/marvell-external-fw-SDK12.25.06/firmware/ebf, Built: Wed, 30 Jul 2025 15:46:57 +0000

Which requires a newer octep_cp_agent version too.

Update to octep_cp_agent version 25.04.0 (commit 35c9be07d2ee). Note
that a newer version of the agent exists, but it seems to cause
problems. We probably will want to update later. For now, update to
25.04.0 only. Actually, 25.04.0 also has problems [1], but still do this
change, as this seems the right direction toward eventually using a
newer firmware.

Note that since the agent is bundled in the VSP, and the required agent
version depends on the firmware version, we would have to update all
systems and the dpu-operator versions at the same time. That makes the
change cumbersome to roll out. Instead, build both variants of the
agent, and run the appropriate one.

In the future, once everybody updated to the newer firmware, we can drop
support for firmware 12.25.01 (and octep_cp_agent 25.03.0).

[1] https://issues.redhat.com/browse/RHEL-107893