marvell: copy vendored "pcie_ep_octeon_target" for building two versions
We will update the firmware version on the Marvell DPUs. But a certain
firmware version also requires a matching cp-agent version. If we just
update the VSP once to use the new cp-agent, then it is cumbersome
because all DPUs and all dpu-operator versions must update at the same
time.

Instead, we will build and bundle two variants of the cp-agent. The new
one, and the old (current) one.

Copy the current version over. This "pcie_ep_octeon_target.25.03.0" is
currently identical to "pcie_ep_octeon_target".

As git tracks files by hashing content, the copied tree is very cheap in
terms of file size in git. This makes the change of copying the entire
vendor dependency less of a problem.

In a next steps, we will build two variants.