marvell: update vendored cp-agent sources
These sources are currently not yet used. They will be in the
next commit.

Next, we will want to build both version 25.03.0 and version 25.04.0
(the version we just reimport here). The benefit is that the VPS will
then be able to handle both firmware versions. That simplies updating
the firmwares, because we can update systems individually.