config: rename default image for network-resources-injector
This container is related to "Dockerfile.networkResourcesInjector.rhel"
Rename the dummy placeholder, so that of the container image resembles
the name of the docker file.

Also, note that due to bug ART-13992, currently we don't build an image
for this docker file yet. We will want to fix that, and then that image
should also have the name "dpu-network-resources-injector".