config: add "MarvellVspCpAgentImage" environment variable
We will need that the container images are configurable via environment
variables in the manifest. That way, ART (the build tooling for
openshift) can patch the manifests with the right container imanges.

That is currently missing for the MarvellCpAgent image. This will need
fixing (https://issues.redhat.com/browse/MDC-107).

In a first step, add the (so far unused) environment variable.