controller: reserve IP range in "default-sriov-net"
Previously, the entire 10.56.217.0/24 was allocatable. That means, there
is no guaranteed set of IP addresses without conflicts.

For example, for a SFC NF we might want to configure static IP addresses
(as we currently don't have a clear IPAM approach yet). By restricting
the range, we can be sure that certain addresses are not automatically
assigned to pods.

The range is chosen somewhat arbitrarily. But it means we have a few IP
addresses on the lower and upper end of the range.

https://issues.redhat.com/browse/IIC-814