Add finalizer to SFC to wait to also delete pods created by SFC
We add a finalizer to the SFC CR such that we can also delete NF pods
that have a label based owner reference to the CR. The
'dpu.sfc.config.openshift.io/owner' label has the namespace and name of
the sfc. Thus when the SFC CR is deleted, it should clean up the
associated NF pod.

This is the technique used to have a owner reference in different
namespaces.

Signed-off-by: William Zhao <wizhao@redhat.com>