manifests: replace tag for image URLs with generic "latest"
These "quay.io" images don't actually exist. They are only a place holder
in the upstream manifest.

When we install the operator from sources, we patch them out anyway (via
the patch in bin/kustomization.yaml and bin/local-images.yaml). When we
build images for Operatorhub in ART, the URLs in the manifest also get
patched out (see [1]). So these names are really only placeholders.

At this point, it is not useful to tie the non-existing tag to the OCP
release. It means when backporting patches we would need to adjust the
patch. And it means when "release-4.20" branch splits from "main"
branch, we need to adjust the files. And also, "release-4.20" and
"release-4.21" currently point to the same git commit where the tag
would not be suitable for "release-4.21" branch.

Instead, use a dummy tag "latest" for all instead of a tag that ties to
the OCP version.

[1] https://github.com/openshift-eng/art-tools/blob/ade5d7adf6b3fb4094f25cdd5dfce96453eb103d/doozer/doozerlib/backend/rebaser.py#L2068