manifests: add image reference for network-resources-injector
ART must onboard the "Dockerfile.networkResourcesInjector.rhel" to build
the "dpu-network-resources-injector" ([1]).

In preparation step for that, fix the dpu-operator side of this.

Note that the image is not yet built, so the renaming of the URL does
not break anything that would work previously. Also, the previous URL
was just wrong.

[1] https://issues.redhat.com/browse/OCPBUGS-60514