Change SFC configuration to use the default namespace
The SFC yaml is most likely to disappear in the future, but the decision
is not made yet. Right now, let's make sure NF pod is outside the
dpu-operator namespace. This is needed for making sure the SFC yaml is
owned by the dpu-operator. And the NF pod is fixed to the default
namespace. This also means the Network Attachment Definition would be
in the default namespace. Why this is needed is because operators should
not have too much administrative capabilities outside it's scope dealing
with Custom Resource Definitions. SFC yaml is part of the operator's
CRD, but the workload should not be owned by the operator. Owner references
on the NF pod are changed to match.

Signed-off-by: William Zhao <wizhao@redhat.com>